+++ src/tests/music-note-fire-player.test.ts
@@ -9,6 +9,7 @@ function fakeAudio(calls: unknown[][]): AudioAdapter {
     isUnlocked: () => true,
     playTone: (...args) => { calls.push(args); },
     playNoise() {},
+    startNoiseLoop: () => ({ stop() {}, setPeak() {}, isPlaying: () => false }),
     setMuted() {},
     isMuted: () => false,
     setVolume() {},